+++ package.json
@@ -79,7 +79,7 @@
       "json",
       "ts"
     ],
-    "rootDir": ".",
+    "rootDir": "test",
     "testRegex": ".spec.ts$",
     "transform": {
       "^.+\\.(t|j)s$": "ts-jest"
