+++ test/command/existing-owner.mock.ts
@@ -1,0 +1,5 @@
+export class ExistingOwnerEventStoreMock {
+    connect = (): void => void 0;
+    exists = (): boolean => true;
+    getEvents = (): [] => [];
+}
