+++ test/command/nonexisting-owner.mock.ts
@@ -1,0 +1,4 @@
+export class NonExistingOwnerEventStoreMock {
+    connect = (): void => void 0;
+    exists = (): boolean => false;
+}
