+++ test/query/controller/sensor.controller.spec.ts
@@ -1,8 +1,8 @@
 import { Test } from '@nestjs/testing';
 import { QueryBus } from '@nestjs/cqrs';
-import { SensorIdParams } from './model/id-params';
-import { SensorController } from './sensor.controller';
-import { RetrieveSensorsParams } from './model/retrieve-sensors-params';
+import { SensorIdParams } from '../../../src/query/controller/model/id-params';
+import { SensorController } from '../../../src/query/controller/sensor.controller';
+import { RetrieveSensorsParams } from '../../../src/query/controller/model/retrieve-sensors-params';
 
 describe('SensorController', () => {
     let queryBus: QueryBus;
